LIQ-211 Fixed stable coin stress test. Combined liquidity actions and… (#87)
* LIQ-211 Fixed stable coin stress test. Combined liquidity actions and swap actions into one test. Re-enabled the stress test

* Cleaned up token allocation in _setupStressTestPool

* Fixed natspec of revenueAvailable

+++ src/amm/mathLibs/MathLibs.sol
@@ -173,10 +173,10 @@ library MathLibs {
         } else {
             if (exponent > decimals) {
                 uint256 diff = uint(exponent - decimals);
-                result = int(uint(mantissa) / (10 ** diff));
+                result = mantissa / int(10 ** diff);
             } else {
                 uint256 diff = uint(decimals - exponent);
-                result = int(uint(mantissa) * (10 ** diff));
+                result = mantissa * int(10 ** diff);
             }
         }
     }
